refactor: CollectionView로 재구성
#93
