+++ Bridge/Sources/Presentation/Common/Extensions/UILabel+.swift
@@ -8,18 +8,6 @@
 import UIKit
 
 extension UILabel {
-    func configureLabel(
-        textColor: UIColor,
-        font: UIFont,
-        textAlignment: NSTextAlignment = .left,
-        numberOfLines: Int = 1
-    ) {
-        self.textColor = textColor
-        self.font = font
-        self.textAlignment = textAlignment
-        self.numberOfLines = numberOfLines
-    }
-    
     // 출처: https://sujinnaljin.medium.com/swift-label의-line-height-설정-및-가운데-정렬-962f7c6e7512
     func configureTextWithLineHeight(text: String?, lineHeight: CGFloat, alignment: NSTextAlignment = .left) {
         guard let text else { return }
@@ -36,6 +24,18 @@ extension UILabel {
         
         let attributedString = NSAttributedString(string: text, attributes: attributes)
         
+        attributedText = attributedString
+    }
+    
+    /// 텍스트의 특정부분에 반전컬러를 적용
+    func highlightTextColor(text: String, highlightText: String) {
+        let attributedString = NSMutableAttributedString(string: text)
+
+        if let rangeOfNumber = text.range(of: highlightText) {
+            let nsRange = NSRange(rangeOfNumber, in: text)
+            attributedString.addAttribute(.foregroundColor, value: BridgeColor.primary1, range: nsRange)
+        }
+
         attributedText = attributedString
     }
 }
