+++ Bridge/Sources/Presentation/ProjectDetail/ProjectDetailViewController.swift
@@ -16,21 +16,6 @@ final class ProjectDetailViewController: BaseViewController {
     // MARK: - UI
     private let rootFlexContainer = UIView()
     
-    private let scrollView: UIScrollView = {
-        let scrollView = UIScrollView()
-        scrollView.backgroundColor = .clear
-        scrollView.showsVerticalScrollIndicator = false
-        
-        return scrollView
-    }()
-    
-    private let contentContainer: UIView = {
-        let view = UIView()
-        view.backgroundColor = BridgeColor.gray09
-        
-        return view
-    }()
-    
     private let dividerView: UIView = {
         let divider = UIView()
         divider.backgroundColor = BridgeColor.gray06
@@ -39,13 +24,27 @@ final class ProjectDetailViewController: BaseViewController {
         return divider
     }()
     
-    private let titleDescriptionView = TitleDescriptionView()
-    private let basicInfoView = BasicInfoView()
-    private let recruitmentFieldInfoView = RecruitmentFieldInfoView()
-
+    private lazy var collectionView: UICollectionView = {
+        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: configureLayout())
+        collectionView.showsVerticalScrollIndicator = false
+        collectionView.register(RecruitFieldCell.self)
+        collectionView.register(
+            ProjectDetailHeaderView.self,
+            forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader
+        )
+        
+        return collectionView
+    }()
+    
     private let menuBar = ProjectDetailMenuBar()
     
     // MARK: - Property
+    private typealias DataSource = UICollectionViewDiffableDataSource<ProjectDetailViewModel.Section, MemberRequirement>
+    private typealias Snapshot = NSDiffableDataSourceSnapshot<ProjectDetailViewModel.Section, MemberRequirement>
+    private var dataSource: DataSource?
+    
+    private var goToDetailButtonTapped = PublishRelay<Void>()
+    
     private let viewModel: ProjectDetailViewModel
     
     // MARK: - Init
@@ -79,52 +78,39 @@ final class ProjectDetailViewController: BaseViewController {
     // MARK: - Layout
     override func configureLayouts() {
         view.addSubview(rootFlexContainer)
-        scrollView.addSubview(contentContainer)
         
         rootFlexContainer.flex.define { flex in
             flex.addItem(dividerView).height(1)
-            flex.addItem().grow(1)
-            
-            flex.addItem(scrollView).position(.absolute).width(100%).top(1).bottom(102)
+            flex.addItem(collectionView).grow(1)
             flex.addItem(menuBar)
-        }
-        
-        contentContainer.flex.define { flex in
-            flex.addItem(titleDescriptionView)
-            flex.addItem(basicInfoView).marginTop(8)
-            flex.addItem(recruitmentFieldInfoView).marginTop(8)
         }
     }
     
     override func viewDidLayoutSubviews() {
         rootFlexContainer.pin.top(view.pin.safeArea.top).left().bottom().right()
         rootFlexContainer.flex.layout()
-        
-        contentContainer.pin.all()
-        contentContainer.flex.layout(mode: .adjustHeight)
-        scrollView.contentSize = contentContainer.frame.size
     }
     
     // MARK: - Binding
     override func bind() {
         let input = ProjectDetailViewModel.Input(
             viewDidLoad: .just(()),
-            goToDetailButtonTapped: recruitmentFieldInfoView.goToDetailButtonTapped
+            goToDetailButtonTapped: goToDetailButtonTapped.asObservable()
         )
         let output = viewModel.transform(input: input)
         
         output.project
-            .drive(onNext: { [weak self] data in
+            .drive(onNext: { [weak self] project in
                 guard let self else { return }
                 
-                self.titleDescriptionView.configureContents(with: data)
-                self.basicInfoView.configureContents(with: data)
-                self.recruitmentFieldInfoView.configureContents(with: data.memberRequirements)
+                self.configureDataSource()
+                self.configureSupplementaryView(with: project)
+                self.applySnapshot(with: project.memberRequirements)
             })
             .disposed(by: disposeBag)
         
         // 구분선 등장
-        scrollView.rx.contentOffset
+        collectionView.rx.contentOffset
             .map { $0.y > 0 }
             .distinctUntilChanged()
             .withUnretained(self)
@@ -132,5 +118,85 @@ final class ProjectDetailViewController: BaseViewController {
                 owner.dividerView.isHidden = !shouldHidden
             })
             .disposed(by: disposeBag)
+    }
+}
+
+// MARK: - CompositionalLayout
+extension ProjectDetailViewController {
+    private func configureLayout() -> UICollectionViewLayout {
+        let itemSize = NSCollectionLayoutSize(
+            widthDimension: .fractionalWidth(1.0),
+            heightDimension: .fractionalHeight(1.0)
+        )
+
+        let item = NSCollectionLayoutItem(layoutSize: itemSize)
+        
+        let groupSize = NSCollectionLayoutSize(
+            widthDimension: .absolute(190),
+            heightDimension: .absolute(127)
+        )
+        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitems: [item])
+        
+        let headerSize = NSCollectionLayoutSize(
+            widthDimension: .fractionalWidth(1),
+            heightDimension: .estimated(640)
+        )
+        
+        let header = NSCollectionLayoutBoundarySupplementaryItem(
+            layoutSize: headerSize,
+            elementKind: UICollectionView.elementKindSectionHeader,
+            alignment: .top
+        )
+        header.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: -16, bottom: 0, trailing: -16)
+        
+        let section = NSCollectionLayoutSection(group: group)
+        section.interGroupSpacing = 8
+        section.boundarySupplementaryItems = [header]
+        section.contentInsets = NSDirectionalEdgeInsets(top: 5, leading: 16, bottom: 20, trailing: 16)
+        section.orthogonalScrollingBehavior = .continuous
+        
+        return UICollectionViewCompositionalLayout(section: section)
+    }
+}
+
+// MARK: - Data source
+extension ProjectDetailViewController {
+    private func configureDataSource() {
+        dataSource = DataSource(
+            collectionView: collectionView
+        ) { collectionView, indexPath, requirement in
+            guard let cell = collectionView.dequeueReusableCell(RecruitFieldCell.self, for: indexPath) else {
+                return UICollectionViewCell()
+            }
+            cell.configureCell(with: requirement)
+            
+            return cell
+        }
+    }
+    
+    private func configureSupplementaryView(with project: Project) {
+        dataSource?.supplementaryViewProvider = { [weak self] collectionView, kind, indexPath in
+            guard let headerView = collectionView.dequeueReusableSupplementaryView(
+                ProjectDetailHeaderView.self,
+                ofKind: kind,
+                for: indexPath
+            ) else { return UICollectionReusableView() }
+            
+            guard let self else { return UICollectionReusableView() }
+            
+            headerView.configureContents(with: project)
+            headerView.goToDetailButtonTapped
+                .bind(to: self.goToDetailButtonTapped)
+                .disposed(by: headerView.disposeBag)
+            
+            return headerView
+        }
+    }
+    
+    private func applySnapshot(with requirements: [MemberRequirement]) {
+        var snapshot = Snapshot()
+        snapshot.appendSections([.main])
+        snapshot.appendItems(requirements)
+        dataSource?.apply(snapshot)
     }
 }
